requirements.txt
streamlit==1.28.1
selenium==4.15.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.3
openpyxl==3.1.2
supabase==2.3.0
python-dotenv==1.0.0
pydantic==2.5.0
